cache password checks to speed up the api
Basic auth does not carry state between requests (by design) so we cannot have a "session ids". Therefor, we are force to validate the password on each and every request. Also by design, kdfs for password hashing are slow and computational intensive.  Its pretty inefficient to run an argon2id hash on every request.
Instead, after the first successful proper validation via argon2, we put a records into an in-memory cache that contains a sha256 hash over the argon2 hash, the users plaintext password, and the IP address sending the request. "Salting" the hash with the IP should make it harder to brute-force a leaked sha256 hash even if the matching argon2 would be known.
We also only keep these records around for  a few minutes after a users latest request. If the users sends another request after the cache entry timed out, we rerun argon2 to verify the password, and add it to the cache again. Requests that hit a cache entry while its still valid extend its lifetime.
We also take some care to zero the memory before dropping the cache entry, to make it harder to extract dropped cache entries.

On my test machine, with the default argon2 settings, this reduces the average duration of a dir listing request from ~40ms to about ~7ms, and even concurrent requests come into play